Move mutation class to graphql_router
